requirements.txt
flask==2.2.2
azure-cosmos==4.0.0